Add files via upload
Introduction
In this project, you need to implement models to predict whether the income of an individual exceeds 50k (labeled as 1) or below 50k (labeled as 0), given the census information of the individual. This dataset is similar to the one in the hw2, but you need to apply the data engineering on your own.

For this project, there are no constraints to the machine learning model. You should try to apply the models we have learned from this course so far.

You will be evaluated over the AUC score on the private leaderboard. Note that you will not see your private leaderboard score before the submission deadline. This competition will close on October 28th.

Suggestions
This dataset contains missing values. Consider impute the data before training the model.
Be aware that the radio of the positive and negative labels are not balanced.